requirements_2.txt
bitsandbytes==0.41.3
dadaptation==3.2
diffusers==0.26.0
ipython==8.15.0
lion_pytorch==0.2.2
lpips==0.1.4
matplotlib==3.9.1
numpy==1.23.5
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
prodigyopt==1.0
pydantic==2.8.2
PyYAML==6.0.2
Requests==2.32.3
safetensors==0.4.4
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
tqdm==4.66.5
transformers==4.26.0
wandb==0.17.0
xformers==0.0.23
accelerate==0.16.0